% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_color_map.R
\name{apply_color_map}
\alias{apply_color_map}
\title{Apply color map to continuous values}
\usage{
apply_color_map(x, n_colors = 1000, color_palette_function = viridis::viridis)
}
\arguments{
\item{x}{data}

\item{n_colors}{number of colors for the palette}

\item{color_palette_function}{callable: color palette function}
}
\value{
color codes for data `x`
}
\description{
Apply color map to continuous values
}
