% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_linear_regression.R
\name{calculate_ml_bayesian_linear_regression}
\alias{calculate_ml_bayesian_linear_regression}
\title{Calculate the marginal likelihood for a Bayesian linear regression model}
\usage{
calculate_ml_bayesian_linear_regression(model, data, return_log = FALSE)
}
\arguments{
\item{model}{List  with model information}

\item{data}{Observed data y}

\item{return_log}{(bool) should the log marginal likelihood be returned}
}
\value{
Marginal Likelihood (log depending on `return_log`)
}
\description{
with Normal-Inverse-Gamma conjugacy
}
