% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_helpers.R
\name{continuity_correction}
\alias{continuity_correction}
\title{Continuity correction}
\usage{
continuity_correction(x, epsilon = 1e-04)
}
\arguments{
\item{x}{a numeric vector}

\item{epsilon}{threshold for truncating}
}
\value{
a numeric vector, corrected
}
\description{
Truncates a vector to the interval (epsilon, 1-epsilon) and re-normalizes the values
}
