% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{flat_lower_triag_to_symmetric}
\alias{flat_lower_triag_to_symmetric}
\title{Build symmetric matrix from flattened lower triangular}
\usage{
flat_lower_triag_to_symmetric(lower_tri, nrow = 2)
}
\arguments{
\item{lower_tri}{flattened lower triangular matrix}

\item{nrow}{number of rows (can be removed with clever deductions in future version)}
}
\value{
constructed symmetric matrix
}
\description{
Build symmetric matrix from flattened lower triangular
}
