% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_helpers.R
\name{log_mean_exp}
\alias{log_mean_exp}
\title{Stable logmeanexp(x)}
\usage{
log_mean_exp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
A numerically stable implementation of log(mean(exp(x))).
}
