% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computational_helpers.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Stable logsumexp(x)}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
A numerically stable implementation of log(sum(exp(x))).
}
