% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_normal_mixture.R
\name{logistic_normal_mixture}
\alias{logistic_normal_mixture}
\title{Construct a logistic normal mixture distribution}
\usage{
logistic_normal_mixture(x, theta, mu_list, Sigma_list, ...)
}
\arguments{
\item{x}{data}

\item{theta}{mixture weights}

\item{mu_list}{(list of) location parameters}

\item{Sigma_list}{(list of) shape parameters}

\item{...}{additional parameters, currently not used}
}
\value{
logistic normal mixture density
}
\description{
Construct a logistic normal mixture distribution
}
