% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simplex_scatter.R
\name{plot_simplex_scatter}
\alias{plot_simplex_scatter}
\title{Plot scatter within a simplex from barycentric coordinates}
\usage{
plot_simplex_scatter(P, ...)
}
\arguments{
\item{P}{matrix of barycentric coordinates}

\item{...}{further arguments passed to `points()`}
}
\description{
Plot scatter within a simplex from barycentric coordinates
}
