% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seir_simulator.R
\name{SEIR}
\alias{SEIR}
\title{Simulate from a basic discrete-time SIR model}
\usage{
SEIR(beta, eta, gamma, N, T_steps, e0 = 1, i0 = 1)
}
\arguments{
\item{beta}{infection rate}

\item{eta}{rate E -> I}

\item{gamma}{recovery rate}

\item{N}{population size}

\item{T_steps}{number of time steps to take}

\item{e0}{number of initially exposed}

\item{i0}{number of initially infected}
}
\value{
I(t) - vector of infected inviduals over T_steps
}
\description{
Simulate from a basic discrete-time SIR model
}
