% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data_and_pmps.R
\name{simulate_data_and_pmps}
\alias{simulate_data_and_pmps}
\title{Simulate set of data and posterior model probabilities}
\usage{
simulate_data_and_pmps(M, prior_model_prob = NULL, K = 1000, N = 20)
}
\arguments{
\item{M}{List of models}

\item{prior_model_prob}{Prior model probabilities, defaults to uniform}

\item{K}{number of simulated data sets}

\item{N}{number of observations per data set}
}
\value{
Data frames with posterior model probabilities and additional info, in wide and long format
}
\description{
Simulate set of data and posterior model probabilities
}
