% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_linear_regression.R
\name{simulate_data_bayesian_linear_regression}
\alias{simulate_data_bayesian_linear_regression}
\title{Data generating process for Bayesian Linear Regression}
\usage{
simulate_data_bayesian_linear_regression(model, N)
}
\arguments{
\item{model}{model details}

\item{N}{number of observations}
}
\value{
data `y`, variances `sigma_2`, weights `beta`
}
\description{
with Normal-Inverse-Gamma Conjugacy
}
