% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_simulator.R
\name{SIR}
\alias{SIR}
\title{Simulate from a basic discrete-time SIR model}
\usage{
SIR(beta, gamma, N, T_steps, i0 = 1)
}
\arguments{
\item{beta}{infection rate}

\item{gamma}{recovery rate}

\item{N}{population size}

\item{T_steps}{number of time steps to take}

\item{i0}{number of initially infected}
}
\value{
I(t) - vector of infected inviduals over T_steps
}
\description{
Simulate from a basic discrete-time SIR model
}
